Move alteration page to a separate component AnnotationPage
